+++ examples/auto_choose_approx.R
@@ -1,31 +1,30 @@
 library(randomForest)
 library(pdp)
 data(boston)
-boston_raw <- boston
 set.seed(101)
 
 # build random forest model:
-boston.rf <- randomForest(cmedv ~ lstat + ptratio + age, data = boston_raw)
+boston.rf <- randomForest(cmedv ~ lstat + ptratio + age, data = boston)
 
-model_pdp <- xp_gam(
+model_pdp <- xspline(
   cmedv ~
-    xs(lstat, spline_opts = list(k = 6), method_opts = list(type = "pdp", grid.resolution = 60)) +
-    xs(ptratio, spline_opts = list(k = 4), method_opts = list(type = "pdp", grid.resolution = 40)) +
+    xs(lstat, transform_opts = list(k = 6), method_opts = list(type = "pdp", grid.resolution = 60)) +
+    xs(ptratio, transform_opts = list(k = 4), method_opts = list(type = "pdp", grid.resolution = 40)) +
     age,
-  blackbox = boston.rf,
-  data = boston_raw
+  model = boston.rf,
+  data = boston
 )
 
-# build xp_gam model with specified response method and approximation options,
+# build xspline model with specified response method and approximation options,
 # also choose automaticaly whether use bare or approximation
-model_pdp_auto <- xp_gam(
+model_pdp_auto <- xspline(
   cmedv ~
-    xs(lstat, spline_opts = list(k = 6), method_opts = list(type = "pdp", grid.resolution = 60)) +
-    xs(ptratio, spline_opts = list(k = 4), method_opts = list(type = "pdp", grid.resolution = 40)) +
+    xs(lstat, transform_opts = list(k = 6), method_opts = list(type = "pdp", grid.resolution = 60)) +
+    xs(ptratio, transform_opts = list(k = 4), method_opts = list(type = "pdp", grid.resolution = 40)) +
     age,
-  blackbox = boston.rf,
-  data = boston_raw,
-  auto_approx = TRUE
+  model = boston.rf,
+  data = boston,
+  alter = list(numeric = 'auto', factor = 'never')
 )
 
 summary(model_pdp)