+++ examples/simple_graphics.R
@@ -6,12 +6,12 @@ set.seed(101)
 # build random forest model:
 boston.rf <- randomForest(cmedv ~ rm + lstat + nox, data = boston)
 
-# build xp_gam model with specified response method and approximation options
-model_pdp <- xp_gam(
+# build xspliner model with specified response method and approximation options
+model_pdp <- xspline(
   cmedv ~ rm + lstat +
-    xs(nox, method_opts = list(type = "pdp", grid.resolution = 100), spline_opts = list(k = 10, bs = "cr")),
-  blackbox = boston.rf
+    xs(nox, method_opts = list(type = "pdp", grid.resolution = 100), transform_opts = list(k = 10, bs = "cr")),
+  model = boston.rf
 )
 
-xp_plot(model_pdp, "nox", data = boston, plot_data = TRUE)
-xp_plot(model_pdp, "nox")
+plot(model_pdp, "nox", data = boston, plot_data = TRUE)
+plot(model_pdp, "nox")