+++ NEWS.md
@@ -1,1 +1,5 @@
-
+# xspliner 0.0.1
+* 
+* Complete package documentation
+* Export only main package functions, non-exported ones moved to utils
+* Basic functionality for xspline.formula method